chore: update dependent packages version
The packages nuxt, husky, lint-staged, and commitlint/cli have been updated to newer versions. The update was needed to maintain good project health, resolve known vulnerabilities in older package versions, and leverage updates and improvements in the recent package releases. The detailed changes for the dependencies and their versions can be found in the package.json file and the pnpm-lock.yaml file.