feat(components): adjust input.vue for better UI alignment
This commit improves UI alignment within the input.vue component. Changes were made to the position of text and padding in several classes to increase readability and aesthetic appeal. Also, the classes for icons got the addition of flexbox, to ensure their center alignment. This provides a more standardized and tidy presentation.