refactor(components): form validation and error handling
This commit refactors the form validation and error handling code in the sign-up and login pages as well as the input component. The changes remove the hard-coded validation messages from the Yup validation schema and replace them with label methods. This makes the error messages more reusable and less verbose, improving the maintainability and readability of the code. Additionally, updated input.vue component to inherit field attributes which provides more control over the validation and error handling process.